resolving type-hinting conflicts
