+++ amqtt/errors.py
@@ -29,11 +29,11 @@ class PluginError(Exception):
 
 
 class PluginImportError(PluginError):
-    def __init__(self, plugin: Any) -> None:
-        super().__init__(f"Plugin import failed: {plugin!r}")
-
+    """Exceptions thrown when loading plugin."""
 
 class PluginInitError(PluginError):
+    """Exceptions thrown when initializing plugin."""
+
     def __init__(self, plugin: Any) -> None:
         super().__init__(f"Plugin init failed: {plugin!r}")
