+++ tests/plugins/mocks.py
@@ -4,10 +4,8 @@ from dataclasses import dataclass
 
 from amqtt.broker import Action
 
-from amqtt.plugins.base import BasePlugin
-from amqtt.plugins.manager import BaseContext
-from amqtt.plugins.topic_checking import BaseTopicPlugin
-from amqtt.plugins.authentication import BaseAuthPlugin
+from amqtt.plugins.base import BasePlugin, BaseAuthPlugin, BaseTopicPlugin
+from amqtt.plugins.contexts import BaseContext
 
 from amqtt.session import Session
