+++ tests/plugins/test_topic_checking.py
@@ -3,8 +3,9 @@ import logging
 import pytest
 
 from amqtt.broker import Action, BrokerContext, Broker
-from amqtt.plugins.manager import BaseContext
-from amqtt.plugins.topic_checking import BaseTopicPlugin, TopicAccessControlListPlugin, TopicTabooPlugin
+from amqtt.plugins.contexts import BaseContext
+from amqtt.plugins.topic_checking import TopicAccessControlListPlugin, TopicTabooPlugin
+from amqtt.plugins.base import BaseTopicPlugin
 from amqtt.session import Session
 
 # Base plug-in object
